+++ src/uma/structural/package_diagram.rs
@@ -10,7 +10,7 @@ impl PackageDiagramExtractor {
     pub fn extract(sta: &SymbolTableArtifact) -> Vec<PackageRecord> {
         let mut packages = Vec::new();
 
-        for sym_id in 0..sta.symbol_count as u32 {
+        for sym_id in 0..sta.symbol_count {
             let sym = match sta.symbol(sym_id) {
                 Some(s) => s,
                 None => continue,
